Housekeeping

See merge request arcanecrypto/lpp!6

+++ internal/platform/ln/ln.go
@@ -21,7 +21,7 @@ import (
 // AddInvoiceData is the data required to add a invoice
 type AddInvoiceData struct {
 	Memo   string `json:"memo"`
-	Amount int64  `json:"amount"`
+	Amount int    `json:"amount"`
 }
 
 // LightningConfig is a struct containing all possible options for configuring
